pyproject.toml
[tool.poetry]
name = "arrival"
version = "0.1.0"
description = ""
authors = ["Nikolay Plyushchay <plushchay.nikolay@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
asyncpg = "^0.25.0"
SQLAlchemy = "^1.4.27"
alembic = "^1.7.5"
envparse = "^0.2.0"
aiohttp = "^3.8.1"
Jinja2 = "^3.0.3"
graphene-sqlalchemy = "^2.3.0"
graphql-server-core = {version = "2.0.0", extras = ["aiohttp"]}
aiohttp-graphql = "1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
